+++ semester2/hw2/task2/arrayStack.h
@@ -6,13 +6,13 @@ class ArrayStack : public Stack
 public:
     ArrayStack(int size = 100);
     ~ArrayStack();
-    void push(int value);
-    int pop();
-    int peek() const;
+    void push(double value);
+    double pop();
+    double peek() const;
     bool isEmpty() const;
     bool isFull() const;
 private:
     int size;
-    int *array;
+    double *array;
     int position;
 };
