+++ semester2/hw2/task2/stack.h
@@ -8,9 +8,9 @@ public:
 
     }
 
-    virtual void push(int value) = 0;
-    virtual int pop() = 0;
-    virtual int peek() const = 0;
+    virtual void push(double value) = 0;
+    virtual double pop() = 0;
+    virtual double peek() const = 0;
     virtual bool isEmpty() const = 0;
     virtual bool isFull() const = 0;
 };
